Fixing out-of-bounds error with some linters
Some linters refer to other files when emitting diagnostics. This is to
be able to provide context for a previous error (e.g. noting where
something was declared when emitting a diagnostic for a mismatched
type).

This change makes it such that if one such diagnostic is emitted, it
won't be displayed as a diagnostic in GitHub, avoiding one out-of-bounds
array access.