Evict the omegaup-specific Python packages (#33)
Currently we need to make a change to this package every single time we
add a new dependency on the omegaup repository. That's a huge drag.

This change expects there to be a virtual environment in the omegaup
repository, so that all the Python tools can be managed over there.